+++ react-client/src/components/FlowBuilder/startingStep.js
@@ -10,20 +10,16 @@ class StartingStep extends React.Component {
 
   render () {
     return (
-      <div style={{border: '1px solid #34bfa3', margin: '0px', borderRadius: '4px'}} className="m-portlet m-portlet--creative m-portlet--bordered-semi">
-        <div className="m-portlet__head">
-          <div className="m-portlet__head-caption">
-            <div className="m-portlet__head-title">
-              <h2 style={{width: '208px', marginLeft: '-30px'}} className="m-portlet__head-label m-portlet__head-label--success">
-                <span style={{textAlign: 'center'}}>
-                  Starting Step
-                </span>
-              </h2>
-            </div>
-          </div>
+      <div style={{borderRadius: '4px'}} className='card'>
+        <div style={{background: '#34bfa3', color: 'white'}} className='card-header'>
+          <h6 style={{marginTop: '10px'}}>
+            Starting Step
+          </h6>
         </div>
-        <div style={{marginTop: '-35px'}} className="m-portlet__body">
-          <COMPONENTSAREA />
+        <div style={{border: '1px solid #34bfa3'}} className='card-body'>
+          <COMPONENTSAREA
+            targetId='starting-step-add-component'
+          />
         </div>
       </div>
     )
